backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
opencv-python
numpy